# core-service/src/main.rs
use axum::{
    response::Json,
    routing::{delete, get, patch, post, put},
    Router,
};
use std::net::SocketAddr;
use tracing_subscriber;
use uuid;

mod ai;
mod config;
mod cors;
mod db;
mod dto;
mod handlers;
mod models;
mod openapi;
mod schema;
mod serde_utils;
mod services;
mod state;

use shared;

#[tokio::main]
async fn main() -> anyhow::Result<()> {
    // 載入 .env 檔案
    dotenv::dotenv().ok();

    // 初始化日誌
    tracing_subscriber::fmt()
        .with_env_filter(
            tracing_subscriber::EnvFilter::from_default_env()
                .add_directive("core_service=debug".parse()?),
        )
        .init();

    tracing::info!("啟動核心服務");

    // 設置數據庫連接池
    let database_url = std::env::var("CORE_DATABASE_URL").unwrap_or_else(|_| {
        "postgresql://bangumi:bangumi_dev_password@172.20.0.2:5432/bangumi".to_string()
    });

    let pool = db::establish_connection_pool(&database_url)?;

    // 嘗試運行遷移（如果 PostgreSQL 未運行將失敗，但這是可以接受的）
    match db::run_migrations(&pool) {
        Ok(_) => tracing::info!("數據庫遷移完成"),
        Err(e) => tracing::warn!("數據庫遷移失敗: {}", e),
    }

    // 初始化應用狀態
    let registry = services::ServiceRegistry::new();
    let app_state = state::AppState::new(pool, registry);

    // 啟動時從資料庫載入已有的所有服務模塊
    load_existing_services(&app_state).await;

    // 啟動 FetchScheduler
    let scheduler = std::sync::Arc::new(services::FetchScheduler::new(app_state.db.clone()));
    let scheduler_clone = scheduler.clone();
    tokio::spawn(async move {
        scheduler_clone.start().await;
    });
    tracing::info!("FetchScheduler started");

    // 啟動 DownloadScheduler
    let download_scheduler = std::sync::Arc::new(services::DownloadScheduler::new(
        app_state.db.clone(),
        app_state.sync_service.clone(),
    ));
    let ds_clone = download_scheduler.clone();
    tokio::spawn(async move {
        ds_clone.start().await;
    });
    tracing::info!("DownloadScheduler started");

    // 啟動 CleanupScheduler
    let cleanup_scheduler =
        std::sync::Arc::new(services::CleanupScheduler::new(app_state.db.clone()));
    let cs_clone = cleanup_scheduler.clone();
    tokio::spawn(async move {
        cs_clone.start().await;
    });
    tracing::info!("CleanupScheduler started");

    // 構建應用路由
    let mut app = Router::new()
        // 服務註冊
        .route("/services/register", post(handlers::services::register))
        .route("/services", get(handlers::services::list_services))
        .route(
            "/services/:service_type",
            get(handlers::services::list_by_type),
        )
        .route(
            "/services/downloader-modules",
            get(handlers::services::list_downloader_modules),
        )
        .route(
            "/services/:service_id/update",
            patch(handlers::services::update_service),
        )
        .route(
            "/services/:service_id/health",
            get(handlers::services::health_check),
        )
        // AnimeWork 管理
        .route("/anime-works", post(handlers::anime::create_anime_work))
        .route("/anime-works", get(handlers::anime::list_anime_work))
        .route(
            "/anime-works/:work_id",
            get(handlers::anime::get_anime_work),
        )
        .route(
            "/anime-works/:work_id",
            axum::routing::delete(handlers::anime::delete_anime_work),
        )
        // Cover 圖片管理
        .route(
            "/anime-works/:work_id/covers",
            axum::routing::get(handlers::covers::list_anime_covers),
        )
        .route(
            "/anime-works/:work_id/covers/:cover_id/set-default",
            axum::routing::post(handlers::covers::set_default_cover),
        )
        // 季度管理
        .route("/seasons", post(handlers::anime::create_season))
        .route("/seasons", get(handlers::anime::list_seasons))
        // Anime 管理（原 anime_series）
        .route("/anime", get(handlers::anime::list_all_anime))
        .route("/anime", post(handlers::anime::create_anime))
        .route(
            "/anime/:anime_id",
            get(handlers::anime::get_anime).put(handlers::anime::update_anime),
        )
        .route(
            "/anime-works/:work_id/anime",
            get(handlers::anime::list_anime),
        )
        // 字幕組管理
        .route(
            "/subtitle-groups",
            post(handlers::anime::create_subtitle_group),
        )
        .route(
            "/subtitle-groups",
            get(handlers::anime::list_subtitle_groups),
        )
        .route(
            "/subtitle-groups/:group_id",
            axum::routing::delete(handlers::anime::delete_subtitle_group),
        )
        // 過濾規則
        .route("/filters", post(handlers::filters::create_filter_rule))
        .route("/filters", get(handlers::filters::get_filter_rules))
        .route("/filters/preview", post(handlers::filters::preview_filter))
        .route(
            "/filters/:rule_id",
            delete(handlers::filters::delete_filter_rule),
        )
        // Webhook 管理
        .route(
            "/webhooks",
            get(handlers::webhooks::list_webhooks).post(handlers::webhooks::create_webhook),
        )
        .route("/webhooks/test", post(handlers::webhooks::test_webhook))
        .route(
            "/webhooks/:id",
            get(handlers::webhooks::get_webhook)
                .put(handlers::webhooks::update_webhook)
                .delete(handlers::webhooks::delete_webhook),
        )
        // 下載記錄
        .route("/downloads", get(handlers::downloads::list_downloads))
        .route("/downloads/retry", post(handlers::downloads::retry_bulk))
        .route(
            "/downloads/:download_id/retry",
            post(handlers::downloads::retry_one),
        )
        // 動畫連結
        .route("/links", post(handlers::links::create_anime_link))
        .route(
            "/links/conflicts",
            get(handlers::links::list_conflicting_links),
        )
        .route("/links/:series_id", get(handlers::links::get_anime_links))
        // 訂閱管理
        .route(
            "/subscriptions",
            post(handlers::subscriptions::create_subscription),
        )
        .route(
            "/subscriptions",
            get(handlers::subscriptions::list_subscriptions),
        )
        .route(
            "/fetcher-modules/:fetcher_id/subscriptions",
            get(handlers::subscriptions::get_fetcher_subscriptions),
        )
        .route(
            "/fetcher-modules",
            get(handlers::subscriptions::list_fetcher_modules),
        )
        .route(
            "/subscriptions/:id",
            delete(handlers::subscriptions::delete_subscription)
                .patch(handlers::subscriptions::update_subscription),
        )
        .route(
            "/subscriptions/:id/fetch",
            post(handlers::subscriptions::trigger_fetch_now),
        )
        .route(
            "/subscriptions/:id/detect-conflicts",
            post(handlers::subscriptions::detect_conflicts_and_generate_filters),
        )
        // Fetcher 結果接收
        .route(
            "/fetcher-results",
            post(handlers::fetcher_results::receive_fetcher_results),
        )
        .route(
            "/raw-fetcher-results",
            post(handlers::fetcher_results::receive_raw_fetcher_results),
        )
        // 解析器管理
        .route(
            "/parsers",
            get(handlers::parsers::list_parsers).post(handlers::parsers::create_parser),
        )
        .route("/parsers/preview", post(handlers::parsers::preview_parser))
        .route(
            "/parsers/:parser_id",
            get(handlers::parsers::get_parser)
                .put(handlers::parsers::update_parser)
                .delete(handlers::parsers::delete_parser),
        )
        // 原始資料管理
        .route("/raw-items", get(handlers::raw_items::list_raw_items))
        .route(
            "/raw-items/count",
            get(handlers::raw_items::count_raw_items),
        )
        .route(
            "/raw-items/:item_id",
            get(handlers::raw_items::get_raw_item),
        )
        .route(
            "/raw-items/:item_id/reparse",
            post(handlers::raw_items::reparse_item),
        )
        .route(
            "/raw-items/:item_id/skip",
            post(handlers::raw_items::skip_item),
        )
        // 訂閱衝突解決
        .route(
            "/conflicts",
            get(handlers::conflict_resolution::get_pending_conflicts),
        )
        .route(
            "/conflicts/:conflict_id/resolve",
            post(handlers::conflict_resolution::resolve_conflict),
        )
        // Anime Link 衝突
        .route(
            "/link-conflicts",
            get(handlers::anime_link_conflicts::list_link_conflicts),
        )
        .route(
            "/link-conflicts/:conflict_id/resolve",
            post(handlers::anime_link_conflicts::resolve_link_conflict),
        )
        .route(
            "/link-conflicts/resolve-by-raw-item",
            post(handlers::anime_link_conflicts::resolve_link_conflicts_by_raw_item),
        )
        // Viewer 同步回呼
        .route("/sync-callback", post(handlers::sync::sync_callback))
        // Dashboard 概覽
        .route(
            "/dashboard/stats",
            get(handlers::dashboard::get_dashboard_stats),
        )
        // 搜尋
        .route("/search", get(handlers::search::search))
        .route("/detail", post(handlers::detail::detail))
        // AI Providers（多 provider 與 fallback chain）
        .route(
            "/ai-providers/reorder",
            post(handlers::ai_providers::reorder_ai_providers),
        )
        .route(
            "/ai-providers/test",
            post(handlers::ai_providers::test_ai_provider_config),
        )
        .route(
            "/ai-providers/:id/test",
            post(handlers::ai_providers::test_ai_provider),
        )
        .route(
            "/ai-providers",
            get(handlers::ai_providers::list_ai_providers)
                .post(handlers::ai_providers::create_ai_provider),
        )
        .route(
            "/ai-providers/:id",
            get(handlers::ai_providers::get_ai_provider)
                .put(handlers::ai_providers::update_ai_provider)
                .delete(handlers::ai_providers::delete_ai_provider),
        )
        // AI Prompt 設定
        .route(
            "/ai-prompt-settings",
            get(handlers::ai_settings::get_ai_prompt_settings)
                .put(handlers::ai_settings::update_ai_prompt_settings),
        )
        .route(
            "/ai-prompt-settings/revert-parser",
            post(handlers::ai_settings::revert_parser_prompt),
        )
        .route(
            "/ai-prompt-settings/revert-filter",
            post(handlers::ai_settings::revert_filter_prompt),
        )
        // 待確認管理
        .route(
            "/pending-ai-results",
            get(handlers::pending_ai_results::list_pending),
        )
        .route(
            "/pending-ai-results/:id",
            get(handlers::pending_ai_results::get_pending)
                .put(handlers::pending_ai_results::update_pending),
        )
        .route(
            "/pending-ai-results/:id/confirm",
            post(handlers::pending_ai_results::confirm_pending),
        )
        .route(
            "/pending-ai-results/:id/reject",
            post(handlers::pending_ai_results::reject_pending),
        )
        .route(
            "/pending-ai-results/:id/regenerate",
            post(handlers::pending_ai_results::regenerate_pending),
        )
        // 健康檢查
        .route("/health", get(health_check))
        // OpenAPI spec
        .route("/openapi.json", get(openapi_json))
        // 應用狀態
        .with_state(app_state);

    // 有條件地應用 CORS 中間件
    if let Some(cors) = cors::create_cors_layer() {
        app = app.layer(cors);
    }

    let port: u16 = std::env::var("SERVICE_PORT")
        .ok()
        .and_then(|s| s.parse().ok())
        .unwrap_or(8000);
    let addr = SocketAddr::from(([0, 0, 0, 0], port));
    let listener = match tokio::net::TcpListener::bind(addr).await {
        Ok(l) => l,
        Err(e) => {
            tracing::error!("無法綁定 {} — {}", addr, e);
            std::process::exit(1);
        }
    };

    tracing::info!("核心服務監聽於 {}", addr);

    axum::serve(listener, app).await?;

    Ok(())
}

async fn health_check() -> Json<serde_json::Value> {
    Json(serde_json::json!({
        "status": "ok",
        "service": "core-service"
    }))
}

async fn openapi_json() -> Json<serde_json::Value> {
    use utoipa::OpenApi;
    let doc = openapi::ApiDoc::openapi();
    Json(serde_json::to_value(doc).unwrap_or_else(|e| {
        tracing::error!("Failed to serialize OpenAPI doc: {}", e);
        serde_json::json!({ "error": "serialization_failed" })
    }))
}

/// Load all service modules (Fetcher, Downloader, Viewer) from database and register them in memory
async fn load_existing_services(app_state: &state::AppState) {
    use crate::models::{ModuleTypeEnum, ServiceModule};
    use crate::schema::service_modules;
    use diesel::prelude::*;

    match app_state.db.get() {
        Ok(mut conn) => {
            // Load all enabled service modules
            match service_modules::table
                .filter(service_modules::is_enabled.eq(true))
                .select(ServiceModule::as_select())
                .load::<ServiceModule>(&mut conn)
            {
                Ok(modules) => {
                    for module in modules {
                        let service_id = uuid::Uuid::new_v4();
                        let (service_type, capabilities) = match module.module_type {
                            ModuleTypeEnum::Fetcher => (
                                shared::ServiceType::Fetcher,
                                shared::Capabilities {
                                    fetch_endpoint: Some("/fetch".to_string()),
                                    search_endpoint: Some("/search".to_string()),
                                    detail_endpoint: Some("/detail".to_string()),
                                    download_endpoint: None,
                                    sync_endpoint: None,
                                    supported_download_types: vec![],
                                },
                            ),
                            ModuleTypeEnum::Downloader => (
                                shared::ServiceType::Downloader,
                                shared::Capabilities {
                                    fetch_endpoint: None,
                                    search_endpoint: None,
                                    detail_endpoint: None,
                                    download_endpoint: Some("/downloads".to_string()),
                                    sync_endpoint: None,
                                    supported_download_types: vec![],
                                },
                            ),
                            ModuleTypeEnum::Viewer => (
                                shared::ServiceType::Viewer,
                                shared::Capabilities {
                                    fetch_endpoint: None,
                                    search_endpoint: None,
                                    detail_endpoint: None,
                                    download_endpoint: None,
                                    sync_endpoint: Some("/sync".to_string()),
                                    supported_download_types: vec![],
                                },
                            ),
                            ModuleTypeEnum::Metadata => (
                                shared::ServiceType::Metadata,
                                shared::Capabilities {
                                    fetch_endpoint: None,
                                    search_endpoint: None,
                                    detail_endpoint: None,
                                    download_endpoint: None,
                                    sync_endpoint: None,
                                    supported_download_types: vec![],
                                },
                            ),
                        };

                        let service = shared::RegisteredService {
                            service_id,
                            service_type: service_type.clone(),
                            service_name: module.name.clone(),
                            host: extract_host(&module.base_url),
                            port: extract_port(&module.base_url),
                            capabilities,
                            is_healthy: true,
                            last_heartbeat: chrono::Utc::now(),
                        };

                        if let Err(e) = app_state.registry.register(service) {
                            tracing::error!(
                                "Failed to load {} {} into registry: {}",
                                module.module_type,
                                module.name,
                                e
                            );
                        } else {
                            tracing::info!(
                                "Loaded {} module from database: {} ({})",
                                module.module_type,
                                module.name,
                                module.base_url
                            );
                        }
                    }
                }
                Err(e) => {
                    tracing::warn!("Failed to load service modules from database: {}", e);
                }
            }
        }
        Err(e) => {
            tracing::error!(
                "Failed to get database connection for loading services: {}",
                e
            );
        }
    }
}

#[cfg(test)]
mod tests {
    use utoipa::OpenApi;

    #[test]
    fn openapi_doc_generates_without_panic() {
        let doc = crate::openapi::ApiDoc::openapi();
        let json = serde_json::to_string(&doc).expect("OpenAPI doc should serialize to JSON");
        assert!(
            json.contains("\"openapi\""),
            "Should contain openapi version field"
        );
        assert!(json.contains("\"paths\""), "Should contain paths field");
        assert!(
            json.contains("/api/core/anime-works"),
            "Should contain anime-works path"
        );
    }

    #[test]
    fn openapi_doc_contains_key_schemas() {
        let doc = crate::openapi::ApiDoc::openapi();
        let json = serde_json::to_string(&doc).expect("OpenAPI doc should serialize to JSON");
        assert!(
            json.contains("AnimeWorkResponse"),
            "Should contain AnimeWorkResponse schema"
        );
        assert!(
            json.contains("AnimeWorkRequest"),
            "Should contain AnimeWorkRequest schema"
        );
        assert!(
            json.contains("DashboardStats"),
            "Should contain DashboardStats schema"
        );
        assert!(
            json.contains("FilterRuleResponse"),
            "Should contain FilterRuleResponse schema"
        );
    }

    #[test]
    fn write_openapi_spec_to_file() {
        let doc = crate::openapi::ApiDoc::openapi();
        let json = serde_json::to_string_pretty(&doc).expect("OpenAPI doc should serialize");
        // 寫入到專案根目錄，供 openapi-typescript 使用
        let out_path = concat!(
            env!("CARGO_MANIFEST_DIR"),
            "/../docs/api/openapi-generated.json"
        );
        std::fs::write(out_path, &json).expect("Should write openapi spec to file");
        println!("Written OpenAPI spec to {}", out_path);
    }

    #[test]
    fn openapi_doc_contains_all_tags() {
        let doc = crate::openapi::ApiDoc::openapi();
        let json = serde_json::to_string(&doc).expect("OpenAPI doc should serialize to JSON");
        for tag in &[
            "AnimeWorks",
            "Seasons",
            "Anime",
            "SubtitleGroups",
            "Filters",
            "Links",
            "Downloads",
            "Dashboard",
        ] {
            assert!(json.contains(tag), "Should contain tag: {}", tag);
        }
    }
}

/// Extract host from URL (e.g., "http://localhost:8001" -> "localhost")
fn extract_host(url: &str) -> String {
    url.split("://")
        .nth(1)
        .and_then(|s| s.split(':').next())
        .unwrap_or("unknown")
        .to_string()
}

/// Extract port from URL (e.g., "http://localhost:8001" -> 8001)
fn extract_port(url: &str) -> u16 {
    url.split(':')
        .last()
        .and_then(|s| s.parse::<u16>().ok())
        .unwrap_or(0)
}