# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: rustBangumi Core Service API
  description: rustBangumi 核心服務 API 規格
  version: 0.4.0
  contact:
    name: rustBangumi Project
  license:
    name: MIT
servers:
  - url: 'http://{host}:{port}'
    description: rustBangumi Core Service
    variables:
      host:
        default: localhost
        description: 主機名稱（Docker 環境請改為 core-service）
      port:
        default: '8000'
        description: 服務端口
...
paths:
...
  /services/register:
    post:
      tags:
        - Services
      summary: 註冊服務
      description: 新的 Fetcher/Downloader/Viewer 服務向核心服務註冊
      operationId: registerService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRegistration'
      responses:
        '201':
          description: 服務註冊成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRegistrationResponse'
        '400':
          description: 請求格式錯誤
...
  /services/{service_type}:
    get:
      tags:
        - Services
      summary: 依類型列出服務
      description: 依服務類型（fetcher/downloader/viewer）列出已註冊的服務
      operationId: listServicesByType
      parameters:
        - name: service_type
          in: path
          required: true
          description: 服務類型
          schema:
            type: string
            enum: [fetcher, downloader, viewer]
      responses:
        '200':
          description: 服務列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: '#/components/schemas/RegisteredService'
...
  /anime-works:
    post:
      tags:
        - AnimeWorks
      summary: 創建動畫作品
      description: 建立一部新的動畫作品記錄
      operationId: createAnimeWork
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnimeWorkRequest'
      responses:
        '201':
          description: 動畫作品創建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimeWorkResponse'
        '500':
          description: 資料庫錯誤

    get:
      tags:
        - AnimeWorks
      summary: 列出所有動畫作品
      description: 獲取所有動畫作品列表
      operationId: listAnimeWorks
      responses:
        '200':
          description: 動畫作品列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  anime_works:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnimeWorkResponse'
  /anime-works/{work_id}:
    get:
      tags:
        - AnimeWorks
      summary: 獲取動畫作品詳情
      description: 依 ID 取得動畫作品詳細資訊
      operationId: getAnimeWork
      parameters:
        - name: work_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 動畫作品詳情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimeWorkResponse'
        '404':
          description: 動畫作品不存在

    delete:
      tags:
        - AnimeWorks
      summary: 刪除動畫作品
      description: 依 ID 刪除動畫作品
      operationId: deleteAnimeWork
      parameters:
        - name: work_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 動畫作品已刪除
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
        '404':
          description: 動畫作品不存在
  /anime-works/{work_id}/covers:
    get:
      tags:
        - AnimeWorks
      summary: 獲取動畫作品封面列表
      description: 依動畫作品 ID 取得所有封面圖列表
      operationId: listAnimeWorkCovers
      parameters:
        - name: work_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 封面列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  covers:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnimeCoverResponse'
        '404':
          description: 動畫作品不存在
  /anime-works/{work_id}/covers/{cover_id}/set-default:
    post:
      tags:
        - AnimeWorks
      summary: 設定預設封面
      description: 將指定封面設為動畫作品的預設封面
      operationId: setDefaultAnimeWorkCover
      parameters:
        - name: work_id
          in: path
          required: true
          schema:
            type: integer
        - name: cover_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 預設封面已更新
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '404':
          description: 動畫作品或封面不存在
  /anime-works/{work_id}/anime:
    get:
      tags:
        - AnimeWorks
      summary: 列出動畫作品的所有動畫
      description: 依動畫作品 ID 列出該作品下的所有動畫
      operationId: listAnimeByAnimeWork
      parameters:
        - name: work_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 動畫列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  anime:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnimeResponse'

  # ============ Seasons ============
  /seasons:
    post:
      tags:
        - Seasons
      summary: 創建季度
      description: 建立一個新的季度記錄
      operationId: createSeason
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSeasonRequest'
      responses:
        '201':
          description: 季度創建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonResponse'
        '500':
          description: 資料庫錯誤

    get:
      tags:
        - Seasons
      summary: 列出所有季度
      description: 獲取所有季度列表
      operationId: listSeasons
      responses:
        '200':
          description: 季度列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  seasons:
                    type: array
                    items:
                      $ref: '#/components/schemas/SeasonResponse'

  # ============ Anime ============
  /anime:
    post:
      tags:
        - Anime
      summary: 創建動畫
      description: 建立一個新的動畫記錄
      operationId: createAnime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnimeRequest'
      responses:
        '201':
          description: 動畫創建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimeResponse'
        '500':
          description: 資料庫錯誤

    get:
      tags:
        - Anime
      summary: 列出所有動畫
      description: 獲取所有動畫列表
      operationId: listAnime
      responses:
        '200':
          description: 動畫列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  anime:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnimeResponse'
  /anime/{anime_id}:
    get:
      tags:
        - Anime
      summary: 依 ID 獲取動畫
      description: 依動畫 ID 取得動畫詳細資訊
      operationId: getAnime
      parameters:
        - name: anime_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 動畫詳情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimeResponse'
        '404':
          description: 動畫不存在

    put:
      tags:
        - Anime
      summary: 更新動畫
      description: 依動畫 ID 更新動畫資訊
      operationId: updateAnime
      parameters:
        - name: anime_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAnimeRequest'
      responses:
        '200':
          description: 動畫已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimeResponse'
        '404':
          description: 動畫不存在

  # ============ Subtitle Groups ============
  /subtitle-groups:
    post:
      tags:
        - SubtitleGroups
      summary: 創建字幕組
      description: 建立一個新的字幕組記錄
      operationId: createSubtitleGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubtitleGroupRequest'
      responses:
        '201':
          description: 字幕組創建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubtitleGroupResponse'
        '500':
          description: 資料庫錯誤

    get:
      tags:
        - SubtitleGroups
      summary: 列出所有字幕組
      description: 獲取所有字幕組列表
      operationId: listSubtitleGroups
      responses:
        '200':
          description: 字幕組列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/SubtitleGroupResponse'
  /subtitle-groups/{group_id}:
    delete:
      tags:
        - SubtitleGroups
      summary: 刪除字幕組
      description: 依 ID 刪除字幕組
      operationId: deleteSubtitleGroup
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 字幕組已刪除
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
        '404':
          description: 字幕組不存在

  # ============ Filter Rules ============
  /filters:
    post:
      tags:
        - Filters
      summary: 創建過濾規則
      description: 建立一條新的過濾規則
      operationId: createFilterRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFilterRuleRequest'
      responses:
        '201':
          description: 過濾規則創建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilterRuleResponse'
        '400':
          description: 請求格式錯誤（target_type 無效或 target_id 設置不正確）

    get:
      tags:
        - Filters
      summary: 獲取過濾規則
      description: 依 target_type 和可選的 target_id 查詢過濾規則，依 rule_order 排序
      operationId: getFilterRules
      parameters:
        - name: target_type
          in: query
          required: true
          description: 目標類型（global/anime-work/anime/group）
          schema:
            type: string
        - name: target_id
          in: query
          required: false
          description: 目標 ID（非全域規則時需指定）
          schema:
            type: integer
      responses:
        '200':
          description: 過濾規則列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      $ref: '#/components/schemas/FilterRuleResponse'
        '400':
          description: target_type 無效
  /filters/{rule_id}:
    delete:
      tags:
        - Filters
      summary: 刪除過濾規則
      description: 依 ID 刪除過濾規則
      operationId: deleteFilterRule
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 過濾規則已刪除
        '404':
          description: 過濾規則不存在

  # ============ Anime Links ============
  /links:
    post:
      tags:
        - Links
      summary: 創建動畫連結
      description: 建立一條新的動畫下載連結
      operationId: createAnimeLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnimeLinkRequest'
      responses:
        '201':
          description: 動畫連結創建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimeLinkResponse'
        '500':
          description: 資料庫錯誤
...
  /raw-items/{item_id}:
    get:
      tags:
        - RawItems
      summary: 獲取原始資料詳情
      description: 依 ID 取得單一原始資料項目
      operationId: getRawItem
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 原始資料詳情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawItemResponse'
        '404':
          description: 項目不存在
...
components:
  schemas:
...
    ServiceRegistration:
      type: object
      required:
        - service_type
        - service_name
        - host
        - port
        - capabilities
      properties:
        service_type:
          type: string
          enum: [fetcher, downloader, viewer]
          description: 服務類型
        service_name:
          type: string
          description: 服務名稱
          example: "mikanani-fetcher"
        host:
          type: string
          description: 服務主機位址
          example: "localhost"
        port:
          type: integer
          description: 服務埠號
          example: 8001
        capabilities:
          $ref: '#/components/schemas/Capabilities'
...
    ServiceRegistrationResponse:
      type: object
      required:
        - service_id
        - registered_at
      properties:
        service_id:
          type: string
          format: uuid
          description: 核心服務分配的服務 UUID
        registered_at:
          type: string
          format: date-time
          description: 註冊時間
...
    CreateAnimeWorkRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: 動畫作品標題
          example: "進擊的巨人"
...